requirements.txt
astroid==2.3.3
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dateparser==0.7.4
docopt==0.6.2
Flask==1.1.2
html5lib==1.0.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
rope==0.16.0
six==1.14.0
soupsieve==2.0
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.8
webencodings==0.5.1
webvtt-py==0.4.5
Werkzeug==1.0.1
wrapt==1.11.2
